requirements.txt
pyautogen==0.2.33
plotly
pandas
torch
kaleido
transformer
uvicorn
fastapi